# src/openapi/api.yaml
openapi: '3.0.3'
info:
  title: API Title
  version: '1.0'
servers:
  - url: http://localhost:3610/v1
...
security:
  - bearerAuth: []
paths:
...
  /users/{userId}:
    get:
      tags:
        - Users 👤
      summary: Retrieve a specific user by their ID
      description: |
        This endpoint retrieves detailed information about a user based on the provided user ID. It returns the user's
        profile data, including their name, email, and biography, if found.

        If the user does not exist, an appropriate error message will be returned.
      x-eov-operation-handler: userController
      operationId: findById
      parameters:
        - name: userId
          in: path
          description: The unique identifier of the user you want to retrieve.
          required: true
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        200:
          description: User successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int64
                    description: HTTP status code indicating successful operation.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the operation.
                    default: OK
                  data:
                    $ref: '#/components/schemas/User'
                additionalProperties: false
                required:
                  - statusCode
                  - message
                  - data
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Users 👤
      summary: Update a user
      description: This endpoint updates the details of an existing user based on the provided user ID.
      x-eov-operation-handler: userController
      operationId: update
      parameters:
        - name: userId
          in: path
          description: The unique identifier of the user to update.
          required: true
          schema:
            type: string
            example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        $ref: '#/components/requestBodies/PutUserRequestBody'
      responses:
        200:
          description: User successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int64
                    description: HTTP status code indicating successful operation.
                    default: 200
                  message:
                    type: string
                    description: Descriptive message of the operation.
                    default: OK
                  data:
                    $ref: '#/components/schemas/User'
                additionalProperties: false
                required:
                  - statusCode
                  - message
                  - data
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Users 👤
      summary: Delete a user by ID
      description: This endpoint deletes a user specified.
      x-eov-operation-handler: userController
      operationId: deleteById
      parameters:
        - name: userId
          in: path
          description: The unique identifier of the user to be deleted.
          required: true
          schema:
            type: string
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        200:
          $ref: '#/components/responses/Ok'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /users/{userId}/likes/{articleId}:
    get:
      tags:
        - Likes 👍
      summary: Check if an article is marked as favorite
      description: Verifies if a specific article is marked as favorite by the user.
      x-eov-operation-handler: likeController
      operationId: checkIfFavorite
      parameters:
        - name: userId
          in: path
          description: The unique identifier of the user.
          required: true
          schema:
            type: string
            format: uuid
            example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        - name: articleId
          in: path
          description: The unique identifier of the article.
          required: true
          schema:
            type: string
            format: uuid
            example: '7cb1d2a5-8c7b-4f3b-a545-2fd938c70c7f'
      responses:
        200:
          description: The article favorite status for the user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  isFavorite:
                    type: boolean
                    description: Indicates if the article is marked as favorite.
                    example: true
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Likes 👍
      summary: Remove a like from an article
      description: Allows a user to remove their like from a specific article using the like ID.
      x-eov-operation-handler: likeController
      operationId: deleteById
      parameters:
        - name: userId
          in: path
          description: The unique identifier of the user.
          required: true
          schema:
            type: string
            format: uuid
            example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        - name: articleId
          in: path
          description: The unique identifier of the article.
          required: true
          schema:
            type: string
            format: uuid
            example: '7cb1d2a5-8c7b-4f3b-a545-2fd938c70c7f'
      responses:
        200:
          $ref: '#/components/responses/Ok'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: |
            The unique identifier for the account, represented as a UUID (Universally Unique Identifier).
            This ID is automatically generated when the account is created.
          example: 550e8400-e29b-41d4-a716-446655440000
        createdAt:
          type: string
          format: date-time
          description: |
            The date and time when the account was created. It is stored in ISO 8601 format (e.g., `2024-10-04T12:35:43.030Z`).
          example: 2024-10-04T12:35:43.030Z
        updatedAt:
          type: string
          format: date-time
          description: |
            The date and time when the account was last updated. It is stored in ISO 8601 format (e.g., `2024-10-04T12:35:43.030Z`).
          example: 2024-10-04T12:35:43.030Z
        name:
          type: string
          description: The name of the user.
          example: John Doe
        email:
          type: string
          description: The email address of the user.
          example: john.doe@example.com
        biography:
          type: string
          description: A short biography of the user.
          example: A software developer from California.
        articles:
          type: integer
          description: The number of the user's articles
          example: 10
        favorites:
          type: integer
          description: The number of the user's favorite articles
          example: 2
      additionalProperties: false
      required:
        - id
        - createdAt
        - updatedAt
        - name
        - email
        - biography
        - articles
        - favorites
...
  requestBodies:
...
    PutUserRequestBody:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
                description: The name of the user.
                example: John Doe
              biography:
                type: string
                description: A short biography of the user.
                example: A software developer from California.
            additionalProperties: false
...
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                format: int64
                description: HTTP status code indicating a successful request.
                default: 200
              message:
                type: string
                description: Descriptive message indicating successful operation.
                default: OK
              data:
                type: object
                description: Additional data returned with the response, if applicable.
            required:
              - statusCode
              - message
            additionalProperties: false
...
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                format: int64
                description: HTTP status code indicating a bad request error.
                default: 400
              message:
                type: string
                description: Descriptive message of the error.
                default: Bad Request
              errorCode:
                type: string
                enum:
                  - BAD_REQUEST
                description: >
                  Possible error codes:

                  - `BAD_REQUEST`: Indicates a bad request error.'
            required:
              - statusCode
              - message
              - errorCode
            additionalProperties: false
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                format: int64
                description: HTTP status code indicating an unauthorized error.
                default: 401
              message:
                type: string
                description: Descriptive message of the error.
                default: Unauthorized
              errorCode:
                type: string
                enum:
                  - UNAUTHORIZED
                description: >
                  Possible error codes:

                  - `UNAUTHORIZED`: Indicates an unauthorized error.'
            required:
              - statusCode
              - message
              - errorCode
            additionalProperties: false
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                format: int64
                description: HTTP status code indicating a forbidden error.
                default: 403
              message:
                type: string
                description: Descriptive message of the error.
                default: Forbidden
              errorCode:
                type: string
                enum:
                  - FORBIDDEN
                description: >
                  Possible error codes:

                  - `FORBIDDEN`: Indicates a forbidden error.'
            required:
              - statusCode
              - message
              - errorCode
            additionalProperties: false
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                format: int64
                description: HTTP status code indicating a not found error.
                default: 404
              message:
                type: string
                description: Descriptive message of the error.
                default: Not Found
              errorCode:
                type: string
                enum:
                  - NOT_FOUND
                  - USER_NOT_FOUND
                description: >
                  Possible error codes:

                  - `NOT_FOUND`: Indicates a not found error.'
            required:
              - statusCode
              - message
              - errorCode
            additionalProperties: false
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                format: int64
                description: HTTP status code indicating an error.
                default: 500
              message:
                type: string
                description: Descriptive message of the error.
                default: Internal Server Error
              errorCode:
                type: string
                enum:
                  - INTERNAL_SERVER_ERROR
                description: >
                  Possible error codes:

                  - `INTERNAL_SERVER_ERROR`: Indicates an error.'
            required:
              - statusCode
              - message
              - errorCode
            additionalProperties: false